+++ config/routes.rb
@@ -10,7 +10,7 @@ Rails.application.routes.draw do
   end
 
   # devise_for :applicants, controllers: { registrations: 'contour/registrations', sessions: 'contour/sessions', passwords: 'contour/passwords' }, path_names: { sign_up: 'register', sign_in: 'login' }
-  devise_for :applicants, path: 'as', path_names: { sign_up: 'register', sign_in: 'login' }
+  devise_for :applicants, path: 'trainee', path_names: { sign_up: 'register', sign_in: 'login' }
 
   resources :applicants do
     collection do
@@ -33,7 +33,7 @@ Rails.application.routes.draw do
     end
   end
 
-  devise_for :preceptors, path: 'ps', path_names: { sign_up: 'register', sign_in: 'login' }
+  devise_for :preceptors, path: 'preceptor', path_names: { sign_up: 'register', sign_in: 'login' }
 
   resources :preceptors do
     collection do
@@ -55,11 +55,13 @@ Rails.application.routes.draw do
     end
   end
 
-  devise_for :users, controllers: { registrations: 'contour/registrations', sessions: 'contour/sessions', passwords: 'contour/passwords', confirmations: 'contour/confirmations', unlocks: 'contour/unlocks' }, path_names: { sign_up: 'register', sign_in: 'login' }
+  devise_for :users, controllers: { registrations: 'contour/registrations', sessions: 'contour/sessions', passwords: 'contour/passwords', confirmations: 'contour/confirmations', unlocks: 'contour/unlocks' }, path_names: { sign_up: 'register', sign_in: 'login' }, path: ""
 
   resources :users
 
   get "/about" => "sites#about", as: :about
+  get "/contact" => "sites#contact", as: :contact
+  get "/forgot-my-password" => "sites#forgot_my_password", as: :forgot_my_password
   get "/settings" => "users#settings", as: :settings
   get "/dashboard" => "sites#dashboard", as: :dashboard