+++ test/integration/registration_test.rb
@@ -4,7 +4,7 @@ class RegistrationTest < ActionDispatch::IntegrationTest
 
   test "a new applicant should be able to sign up" do
     assert_difference('Applicant.count') do
-      post "/as", applicant: { first_name: 'First Name', last_name: 'Last Name', email: 'new_applicant@example.com', password: 'password', password_confirmation: 'password' }
+      post "/trainee", applicant: { first_name: 'First Name', last_name: 'Last Name', email: 'new_applicant@example.com', password: 'password', password_confirmation: 'password' }
     end
 
     assert_not_nil assigns(:applicant)